backend/requirements.txt
fastapi==0.136.0
uvicorn==0.44.0
SQLAlchemy==2.0.49
pydantic==2.13.2
pydantic-settings==2.13.1
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.26
email-validator==2.3.0
python-dotenv==1.2.2

pypdf==6.10.2
python-docx==1.2.0

spacy==3.8.13
sentence-transformers==5.1.2
torch

numpy
scikit-learn